requirements.txt
pyyaml==5.4.1
prometheus_client==0.17.1
GitPython==3.1.16
